v1.13.7: Fix all clippy warnings across codebase
Resolved 21 clippy warnings: simplified redundant closures, replaced
manual flatten with .flatten(), used derive(Default) where possible,
removed unnecessary casts, and added allow attributes for functions
with many arguments (Tauri commands with multiple params).

+++ src-tauri/src/wabbajack_downloader.rs
@@ -346,7 +346,7 @@ impl WjDownloader {
             WjProgressEvent::DownloadProgress {
                 archive_name: archive_name.to_string(),
                 bytes_downloaded: 0,
-                total_bytes: file_node.size() as u64,
+                total_bytes: file_node.size(),
             },
         );
